+++ test/test_retry.rb
@@ -112,6 +112,7 @@ class TestRetry < MiniTest::Unit::TestCase
     it 'should poll like a bad mother...SHUT YO MOUTH' do
       fake_msg = Sidekiq.dump_json({ 'class' => 'Bob', 'args' => [1,2], 'queue' => 'someq' })
       @redis.expect :multi, [[fake_msg], 1], []
+      @redis.expect :multi, [[], nil], []
       @redis.expect :rpush, 1, ['queue:someq', fake_msg]
 
       inst = Sidekiq::Scheduled::Poller.new